+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Improvements:
 * [OLMIS-4795](https://openlmis.atlassian.net/browse/OLMIS-4795): Updated dev-ui to version 8.
 * [OLMIS-4813](https://openlmis.atlassian.net/browse/OLMIS-4813): Updated datepickers to use the new syntax.
 * [OLMIS-4813](https://openlmis.atlassian.net/browse/OLMIS-4813): Updated ui-components to version 6.0.0.
+* [OLMIS-4535](https://openlmis.atlassian.net/browse/OLMIS-4535): Added an abstract state for inventory item details and status update modals.
 
 Bug fixes:
 * [OLMIS-4910](https://openlmis.atlassian.net/browse/OLMIS-4910): Fixed broken inventory list when there are no inventory items.